+++ crates/client/src/content.rs
@@ -88,6 +88,7 @@ impl Content {
             // The returned URL is relative to the learn base, and is normally broken and shows the old learn interface nested a bunch of times
             // This is fixed by adding `&from_ultra=true`, as learn ultra does.
             Some(ContentDetail::Piazza { launch_link }) => ContentPayload::Piazza(format!("{}{}&from_ultra=true", LEARN_BASE, launch_link)),
+            Some(ContentDetail::MediaHopperReplay { launch_link }) => ContentPayload::MediaHopperReplay(format!("{}{}&from_ultra=true", LEARN_BASE, launch_link)),
             Some(ContentDetail::Unknown {}) | None => ContentPayload::Other,
         };
 
@@ -112,7 +113,8 @@ impl Content {
         match &self.payload {
             ContentPayload::Link(link) => link,
             ContentPayload::File { permanent_url, .. } => permanent_url,
-            ContentPayload::Piazza(u) => u,
+            ContentPayload::Piazza(u)
+            | ContentPayload::MediaHopperReplay(u) => u,
             _ => &self.link,
         }
     }
@@ -142,6 +144,9 @@ pub enum ContentPayload {
 
     /// Link to a piazza forum. Payload is the link to open to authenticate then redirect to it.
     Piazza(String),
+
+    /// Link to a media hopper replay class. Payload is the link to open to authenticate then redirect to it.
+    MediaHopperReplay(String),
 }
 
 #[derive(Deserialize)]
@@ -200,6 +205,10 @@ enum ContentDetail {
     #[serde(rename = "resource/x-bb-bltiplacement-49f1179af0494f078ce3ff737dd75de4")]
     #[serde(rename_all = "camelCase")]
     Piazza { launch_link: String },
+
+    #[serde(rename = "resource/x-bb-bltiplacement-mhrlti")]
+    #[serde(rename_all = "camelCase")]
+    MediaHopperReplay { launch_link: String },
 
     #[serde(untagged)]
     Unknown {},